fix: more startup problems

+++ src/main/java/me/jellysquid/mods/lithium/mixin/chunk/section_update_tracking/ChunkHolderMixin.java
@@ -32,7 +32,7 @@ public class ChunkHolderMixin {
                     shift = At.Shift.BEFORE
             )
     )
-    private byte createShortHashSet(byte b) {
+    private int createShortHashSet(int b) {
         if (blockUpdatesBySection[b] == null) {
             this.pendingBlockUpdates = true;
             this.blockUpdatesBySection[b] = new ShortOpenHashSet();
